requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-cors-headers==2.5.2
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
enum34==1.1.6
idna==2.8
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.9
termcolor==1.1.0
urllib3==1.24.1
vk-api==11.4.0
xlrd
